requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
google-generativeai>=0.4.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-pptx>=0.6.21
faiss-cpu>=1.7.4
networkx>=3.1
pyvis>=0.3.2
matplotlib>=3.7.2
beautifulsoup4>=4.12.2
tiktoken>=0.5.1
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
pillow>=10.0.0
protobuf>=4.24.4
plotly>=5.18.0
sympy>=1.12
mammoth>=1.6.0
